refactor(ipc): modularize ipc.ts by IpcChannel prefix
Split the monolithic 793-line ipc.ts into 12 focused modules under
src/main/ipc/ based on IpcChannel prefix groups:
- app, system, file, fileUtils, backup, knowledge, auth,
  agents, integrations, notification, utility, and index orchestrator.

No behavioral changes — pure structural refactor.
