+++ src/main/ipc/utility.ts
@@ -1,0 +1,19 @@
+import { IpcChannel } from '@shared/IpcChannel'
+import { ipcMain } from 'electron'
+
+import { decrypt, encrypt } from '../utils/aes'
+import { compress, decompress } from '../utils/zip'
+
+export function registerUtilityIpc() {
+  // zip
+  ipcMain.handle(IpcChannel.Zip_Compress, (_, text: string) => compress(text))
+  ipcMain.handle(IpcChannel.Zip_Decompress, (_, text: Buffer) => decompress(text))
+
+  // aes
+  ipcMain.handle(IpcChannel.Aes_Encrypt, (_, text: string, secretKey: string, iv: string) =>
+    encrypt(text, secretKey, iv)
+  )
+  ipcMain.handle(IpcChannel.Aes_Decrypt, (_, encryptedData: string, iv: string, secretKey: string) =>
+    decrypt(encryptedData, iv, secretKey)
+  )
+}
